Add thread UI language helper for tests
