+++ tests/helpers/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "test_helpers"
+version = "0.0.0"
+edition = "2018"
+authors = ["Rafael Rivera", "Microsoft"]
+publish = false
+
+[dependencies]
+windows = { path = "../.." }
+
+[build-dependencies]
+windows = { path = "../.." }
