+++ tests/helpers/build.rs
@@ -1,0 +1,7 @@
+fn main() {
+    windows::build! {
+        Windows::Win32::Globalization::{
+            SetThreadPreferredUILanguages, MUI_LANGUAGE_NAME,
+        },
+    };
+}
